requirements.txt
aiogram<3
notion-client
python-dotenv